Refactor to have common method to filter compatible recommendations
