+++ client/src/components/Workflow/Editor/Recommendations.vue
@@ -19,6 +19,7 @@
 
 <script>
 import { getToolPredictions } from "./modules/services";
+import { getCompatibleRecommendations } from "./modules/utilities";
 import LoadingSpan from "components/LoadingSpan";
 import { toSimple } from "./modules/model";
 import _l from "utils/localization";
@@ -120,38 +121,14 @@ export default {
                 this.isDeprecated = predictedData.is_deprecated;
                 this.deprecatedMessage = predictedData.message;
                 if (predictedDataChildren.length > 0) {
-                    const cTools = [];
-                    const toolMap = new Map();
-                    for (const nameObj of predictedDataChildren.entries()) {
-                        const inputDatatypes = nameObj[1].i_extensions;
-                        for (const outT of outputDatatypes.entries()) {
-                            for (const inTool of inputDatatypes.entries()) {
-                                if (
-                                    this._isSubType(outT[1], inTool[1]) ||
-                                    outT[1] === "input" ||
-                                    outT[1] === "_sniff_" ||
-                                    outT[1] === "input_collection"
-                                ) {
-                                    const toolId = nameObj[1].tool_id;
-                                    if (!toolMap.has(toolId)) {
-                                        toolMap.set(toolId, true);
-                                        cTools.push({
-                                            id: toolId,
-                                            name: nameObj[1].name
-                                        });
-                                        break;
-                                    }
-                                }
-                            }
-                        }
-                    }
-                    this.compatibleTools = cTools;
+                    this.compatibleTools = getCompatibleRecommendations(
+                        predictedDataChildren,
+                        outputDatatypes,
+                        this.datatypesMapper
+                    );
                 }
                 this.showLoading = false;
             });
-        },
-        _isSubType(child, parent) {
-            return this.datatypesMapper.isSubType(child, parent);
         },
     },
 };
