+++ client/src/components/Workflow/Editor/modules/utilities.js
@@ -203,3 +203,32 @@ export function saveAs(workflow) {
         Cancel: hide_modal,
     });
 }
+
+export function getCompatibleRecommendations(predChild, outputDatatypes, datatypesMapper) {
+    const cTools = [];
+    const toolMap = new Map();
+    for (const nameObj of predChild.entries()) {
+        const inputDatatypes = nameObj[1].i_extensions;
+        for (const outT of outputDatatypes.entries()) {
+            for (const inTool of inputDatatypes.entries()) {
+                if (
+                    datatypesMapper.isSubType(outT[1], inTool[1]) ||
+                    outT[1] === "input" ||
+                    outT[1] === "_sniff_" ||
+                    outT[1] === "input_collection"
+                ) {
+                    const toolId = nameObj[1].tool_id;
+                    if (!toolMap.has(toolId)) {
+                        toolMap.set(toolId, true);
+                        cTools.push({
+                            id: toolId,
+                            name: nameObj[1].name,
+                        });
+                        break;
+                    }
+                }
+            }
+        }
+    }
+    return cTools;
+}
